+++ src/component/layout/defaultLayout.tsx
@@ -25,177 +25,36 @@ import Paper from "@mui/material/Paper";
 import { styled } from "@mui/material/styles";
 import { PrimaryButton } from "../button/Button";
 import AccordionMenu from "../accordion/AccordionMenu";
-
-const pages = [
-  { label: "Top", href: "/" },
-  { label: "About", href: "/#about" },
-  { label: "Works", href: "/#work" },
-  { label: "Blog", href: "/blog" },
-];
-const settings = ["Profile", "Account", "Dashboard", "Logout"];
-
+import Header from "./Header";
+import Footer from "./Footer";
 function DefaultLayout({ children }: { children: any }) {
-  const [anchorElNav, setAnchorElNav] = React.useState<null | HTMLElement>(
-    null
-  );
-  const [anchorElUser, setAnchorElUser] = React.useState<null | HTMLElement>(
-    null
-  );
-
-  const handleOpenNavMenu = (event: React.MouseEvent<HTMLElement>) => {
-    setAnchorElNav(event.currentTarget);
-  };
-  const handleOpenUserMenu = (event: React.MouseEvent<HTMLElement>) => {
-    setAnchorElUser(event.currentTarget);
-  };
-
-  const handleCloseNavMenu = () => {
-    setAnchorElNav(null);
-  };
-
-  const handleCloseUserMenu = () => {
-    setAnchorElUser(null);
-  };
-
-  const styles = {
-    appBar: {
-      background: "linear-gradient(95deg, #ffc97a, #ff9900)",
-    },
-  };
-  const Item = styled(Paper)(({ theme }) => ({
-    backgroundColor: theme.palette.mode === "dark" ? "#1A2027" : "#fff",
-    ...theme.typography.body2,
-    padding: theme.spacing(1),
-    textAlign: "center",
-    color: theme.palette.text.secondary,
-  }));
   return (
     <React.Fragment>
-      <Paper sx={{ p: 2, display: "flex", justifyContent: "space-between" }}>
-        <Typography>こんにちは〇〇さん</Typography>
-        <PrimaryButton
-          sx={{
-            background: "linear-gradient(95deg, #ffc97a, #ff9900)",
-            borderRadius: 4,
-          }}
-        >
-          管理者用TOP
-        </PrimaryButton>
-      </Paper>
-      <AppBar position="static" sx={styles.appBar}>
-        <Container maxWidth="xl">
-          <Toolbar disableGutters>
-            <Box sx={{ flexGrow: 1, display: { xs: "flex", md: "none" } }}>
-              <IconButton
-                size="large"
-                aria-label="account of current user"
-                aria-controls="menu-appbar"
-                aria-haspopup="true"
-                onClick={handleOpenNavMenu}
-                color="inherit"
-              >
-                <MenuIcon />
-              </IconButton>
-              <Menu
-                id="menu-appbar"
-                anchorEl={anchorElNav}
-                anchorOrigin={{
-                  vertical: "bottom",
-                  horizontal: "left",
-                }}
-                keepMounted
-                transformOrigin={{
-                  vertical: "top",
-                  horizontal: "left",
-                }}
-                open={Boolean(anchorElNav)}
-                onClose={handleCloseNavMenu}
-                sx={{
-                  display: { xs: "block", md: "none" },
-                }}
-              >
-                {pages.map((page) => (
-                  <MenuItem key={page.label} onClick={handleCloseNavMenu}>
-                    <Typography textAlign="center">{page.label}</Typography>
-                  </MenuItem>
-                ))}
-              </Menu>
-            </Box>
-            <AdbIcon sx={{ display: { xs: "flex", md: "none" }, mr: 1 }} />
-            <Typography
-              variant="h5"
-              noWrap
-              component="a"
-              href=""
-              sx={{
-                mr: 2,
-                display: { xs: "flex", md: "none" },
-                flexGrow: 1,
-                fontFamily: "monospace",
-                fontWeight: 700,
-                letterSpacing: ".3rem",
-                color: "inherit",
-                textDecoration: "none",
-              }}
-            >
-              LOGO
-            </Typography>
-            <Box sx={{ flexGrow: 1, display: { xs: "none", md: "flex" } }}>
-              {pages.map((page) => (
-                <Button
-                  key={page.label}
-                  href={page.href}
-                  onClick={handleCloseNavMenu}
-                  sx={{ my: 2, color: "white", display: "block" }}
-                >
-                  {page.label}
-                </Button>
-              ))}
-            </Box>
-
-            <Box sx={{ flexGrow: 0 }}>
-              <Menu
-                sx={{ mt: "45px" }}
-                id="menu-appbar"
-                anchorEl={anchorElUser}
-                anchorOrigin={{
-                  vertical: "top",
-                  horizontal: "right",
-                }}
-                keepMounted
-                transformOrigin={{
-                  vertical: "top",
-                  horizontal: "right",
-                }}
-                open={Boolean(anchorElUser)}
-                onClose={handleCloseUserMenu}
-              >
-                {settings.map((setting) => (
-                  <MenuItem key={setting} onClick={handleCloseUserMenu}>
-                    <Typography textAlign="center">{setting}</Typography>
-                  </MenuItem>
-                ))}
-              </Menu>
-            </Box>
-          </Toolbar>
-        </Container>
-      </AppBar>
+      <Header />
 
       <Container maxWidth="xl" sx={{ my: 20 }}>
         <Grid container spacing={2}>
           <Grid item xs={3}>
             <AccordionMenu />
-            <Box>
-              <TextField />
-            </Box>
+            {/* 検索ボタン */}
+            <Box sx={{ my: 10 }}>
+              <TextField
+                id="outlined-basic"
+                variant="outlined"
+                sx={{ my: 4, width: "100%", backgroundColor: "#fff" }}
+              />
 
-            <PrimaryButton
-              sx={{
-                background: "linear-gradient(95deg, #ffc97a, #ff9900)",
-              }}
-            >
-              検索する
-            </PrimaryButton>
+              <PrimaryButton
+                sx={{
+                  background: "linear-gradient(95deg, #ffc97a, #ff9900)",
+                  mb: 4,
+                  width: "100%",
+                  fontWeight: "bold",
+                }}
+              >
+                検索する
+              </PrimaryButton>
+            </Box>
           </Grid>
 
           <Grid item xs={9}>
@@ -204,11 +63,8 @@ function DefaultLayout({ children }: { children: any }) {
           </Grid>
         </Grid>
       </Container>
-      <Paper sx={{ display: "flex" }}>
-        <Typography>ご利用ガイド</Typography>
-        <Typography>お問い合わせ</Typography>
-        <Typography>投票</Typography>
-      </Paper>
+
+      <Footer />
     </React.Fragment>
   );
 }
